PIN-3695: Rename files
